+++ examples/vite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # dde-earth-vite-example
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @dde-earth/plugin-tiff-loader@1.0.4
+
 ## 0.0.3
 
 ### Patch Changes