+++ packages/plugin-tiff-loader/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dde-earth/plugin-tiff-loader
 
+## 1.0.4
+
+### Patch Changes
+
+- fix tiff render error
+
 ## 1.0.3
 
 ### Patch Changes